feat: transition 제거
